Enhancement: Merge MaximumDuration into Duration
